requirements.txt
### Requirements for OVERWATCH ###
# Required

# General
enum34

# For Python 2/3 support
future

# Flask
flask
Flask-Login
Flask-Assets
Flask-RESTful

# Database
ZODB
#Flask-ZODB
# Install from git repo to support the newer hook
# git+https://github.com/dag/flask-zodb.git
# for compability for python 3
git+https://github.com/SpotlightKid/flask-zodb.git

# Login
bcrypt
Flask-Bcrypt

# Status page
requests

# WSGI
flup6
# Not necessarily required
uwsgi

# Processing
numpy
# ROOT (required, but cannot be installed by pip)

# Documentation
sphinx
sphinxcontrib-napoleon
recommonmark
pprint

# Configuration
pyyaml
ruamel.yaml<=0.15
